+++ src/methods/$cycle/__tests__/$cycle.test.js
@@ -1,44 +1,28 @@
-import { $, $async, $await, $iteratorSymbol } from '../../../../generate/async.macro';
+import { $, $async, $await } from '../../../../generate/async.macro';
 
-import { $cycle, range } from '../../..';
+import { $cycle, $slice, $toArray, $wrap, range } from '../../..';
 
 describe($`cycle`, () => {
   it(
     'return infinite cycle',
     $async(() => {
-      const iter = $cycle([1, 2, 3])[$iteratorSymbol]();
-      expect($await(iter.next())).toEqual({ value: 1, done: false });
-      expect($await(iter.next())).toEqual({ value: 2, done: false });
-      expect($await(iter.next())).toEqual({ value: 3, done: false });
-      expect($await(iter.next())).toEqual({ value: 1, done: false });
-      expect($await(iter.next())).toEqual({ value: 2, done: false });
-      expect($await(iter.next())).toEqual({ value: 3, done: false });
-    }),
-  );
-
-  it(
-    'return infinite cycle (from iterator)',
-    $async(() => {
-      const iter = $cycle(range(3))[$iteratorSymbol]();
-      expect($await(iter.next())).toEqual({ value: 0, done: false });
-      expect($await(iter.next())).toEqual({ value: 1, done: false });
-      expect($await(iter.next())).toEqual({ value: 2, done: false });
-      expect($await(iter.next())).toEqual({ value: 0, done: false });
-      expect($await(iter.next())).toEqual({ value: 1, done: false });
-      expect($await(iter.next())).toEqual({ value: 2, done: false });
+      expect($await($toArray($slice(0, 6, $cycle($wrap([1, 2, 3])))))).toEqual([1, 2, 3, 1, 2, 3]);
     }),
   );
 
   it(
     'can be reused',
     $async(() => {
-      const myCycle = $cycle(range(3));
-      const iter1 = myCycle[$iteratorSymbol]();
-      expect($await(iter1.next())).toEqual({ value: 0, done: false });
-      expect($await(iter1.next())).toEqual({ value: 1, done: false });
-      const iter2 = myCycle[$iteratorSymbol]();
-      expect($await(iter2.next())).toEqual({ value: 0, done: false });
-      expect($await(iter2.next())).toEqual({ value: 1, done: false });
+      const myCycle = $cycle(range(1, 4));
+      expect($await($toArray($slice(0, 7, myCycle)))).toEqual([1, 2, 3, 1, 2, 3, 1]);
+      expect($await($toArray($slice(0, 7, myCycle)))).toEqual([1, 2, 3, 1, 2, 3, 1]);
+    }),
+  );
+
+  it(
+    'can cycle a limited number of times',
+    $async(() => {
+      expect($await($toArray($cycle(3, $wrap([1, 2, 3]))))).toEqual([1, 2, 3, 1, 2, 3, 1, 2, 3]);
     }),
   );
 });