+++ src/methods/$cycle/async-cycle.mjs
@@ -1,10 +1,16 @@
-import { asyncWrapWithMethodIterable } from '../../internal/async-iterable';
+import { asyncIterableCurry } from '../../internal/async-iterable';
 
 import { asyncToArray } from '../$to-array/async-to-array';
 import { cycle } from './cycle';
 
-export async function* asyncCycle(iterable) {
-  yield* cycle(await asyncToArray(iterable));
+export async function* asyncCycle(iterable, times = Infinity) {
+  yield* cycle(await iterable, times);
 }
 
-export default asyncWrapWithMethodIterable(asyncCycle);
+export default asyncIterableCurry(asyncCycle, {
+  minArgs: 0,
+  maxArgs: 1,
+  validateArgs(args) {
+    args[1] = asyncToArray(args[1]);
+  },
+});