+++ src/methods/$take-sorted/$take-sorted.d.ts
@@ -7,15 +7,24 @@ declare function $takeSorted<T = any>(
 ): (source: $SourceIterable<T>) => $ResultIterable<T>;
 
 declare function $takeSorted<T = any>(
-  func: (a: T, b: T) => number,
   n: number,
+  func: (a: T, b: T) => number,
+): (source: $SourceIterable<T>) => $ResultIterable<T>;
+
+declare function $takeSorted<T = any>(
+  func: (a: T, b: T) => number,
 ): (source: $SourceIterable<T>) => $ResultIterable<T>;
 
 declare function $takeSorted<T = any>(n: number, source: $SourceIterable<T>): $ResultIterable<T>;
 
 declare function $takeSorted<T = any>(
-  func: (a: T, b: T) => number,
   n: number,
+  func: (a: T, b: T) => number,
+  source: $SourceIterable<T>,
+): $ResultIterable<T>;
+
+declare function $takeSorted<T = any>(
+  func: (a: T, b: T) => number,
   source: $SourceIterable<T>,
 ): $ResultIterable<T>;