+++ src/methods/$window/async-window.mjs
@@ -15,7 +15,7 @@ export async function* asyncWindow(source, size, { filler } = {}) {
   circular.fill(filler);
   let index = 0;
 
-  for await (const item of asyncConcat(source, repeat(filler, size - 1))) {
+  for await (const item of asyncConcat(source, repeat(size - 1, filler))) {
     circular.push(item);
 
     if (index + 1 >= size) {