+++ src/methods/repeat/repeat.mjs
@@ -1,9 +1,19 @@
 import { wrapWithMethodIterable } from '../../internal/iterable';
 
-export function* repeat(item, times = Infinity) {
-  while (times--) {
+export function* repeat(n, item) {
+  while (n--) {
     yield item;
   }
 }
 
-export default wrapWithMethodIterable(repeat);
+export default wrapWithMethodIterable(repeat, {
+  validateArgs(args) {
+    if (args.length === 0) {
+      throw new Error('Required `item` argument to `repeat` not specified');
+    } else if (args.length === 1) {
+      args.unshift(Infinity);
+    } else if (args[0] === undefined) {
+      args[0] = Infinity;
+    }
+  },
+});